changed consts on UserScriptEditor
