+++ packages/suite-base/src/players/UserScriptPlayer/transformerWorker/typescript/errors.ts
@@ -17,7 +17,7 @@ import {
   DiagnosticSeverity,
   Sources,
   ErrorCodes,
-} from "@lichtblick/suite-base/players/UserScriptPlayer/types";
+} from "@lichtblick/suite-base/players/UserScriptPlayer/constants";
 
 export const noFuncError = {
   severity: DiagnosticSeverity.Error,
